fix: ensure email retrieval and order processing in EventHandler
This commit addresses two main issues in `EventHandler.php`.
Firstly, the fetching of the email from the `GuestOrder` session is moved at a different location
in the flow where it makes more sense and doesn't get overridden prematurely.

Secondly, an exhaustive condition handling has been added to address cases where there might be a
payment error or async payment error. The method fetches the current order via the URL hash which
could be an `OrderInProcess` or an already existing `Order`.

The commit also ensures that the customer UUID is set correctly if the session exists. An update is
made to the database to accommodate this change.

Minor changes have also been made to spacing and data-type declaration in `GuestOrderUser.php` to
help with readability and consistency.

Related: quiqqer/order-guestorder#11