fix: update order processing in eventhandler
This commit addresses the event handling within the QUI ERP order process. Instead of returning the
order directly, the changes allow update of the 'c_user' field in the database by using the
customer UUID from the session. Additionally, it now returns the order fetched by hash instead of
the direct order object. This approach is more fail-safe and consistent.

Related: quiqqer/order-guestorder#11