feat: add new events for simple checkout process
This commit introduces two new events, `onQuiqqerSimpleCheckoutBodyEnd` and
`onQuiqqerSimpleCheckoutValidation`. The functions associated with these events handle the logic of
the checkout process for guests, specifically concerning the display and validation of delivery,
shipping, and billing addresses based on whether the order is anonymous and does not exceed a
certain sum. The total sum is retrieved from the configuration of the `quiqqer/order-guestorder`
package.