fix: improve email attribute handling in guest order user
Made two significant changes to the 'GuestOrderUser.php' file, all aimed at enhancing the handling
for the email attribute.

Firstly, switched from not operator (!) to empty for checking the 'email' attribute, this should
now catch cases where 'email' is null or an empty string.

Second, added the user email to the created address instance and handled possible exceptions. This
ensures the 'email' attribute of the user is also part of the address instance, increasing data
consistency.

Related: quiqqer/order#180