fix: improve guest order user email handling
This commit refactors the `GuestOrderUser.php` to enhance email attribute handling. A new method
`readEmail()` is introduced to isolate the process of reading a guest user's email. This method is
called during the process of constructing the `GuestOrderUser` and also during the creation of a
new `QUI\ERP\Address` instance. On failing to add email to the `Address` instance, an error message
is logged using `QUI\System\Log::addError`. This should help in better debugging and maintaining
the quality of the application.

Related: quiqqer/order#180